# Faster large screen updates + scrolling detection