--- a/module/Makefile.am
+++ b/module/Makefile.am
@@ -1,5 +1,5 @@
 EXTRA_DIST = common.asm nasm_lt.sh
-EXTRA_FLAGS =
+EXTRA_FLAGS = --std=gnu90 --pedantic
 EXTRA_SOURCES =
 EXTRA_HEADERS =
 SUBDIRS =