… module/rdpCapture.h
… module/rdpClientCon.c
… module/rdpPutImage.c
… module/rdpRandR.c
… module/wyhash.h