This patch switches from slow byte-serial CRC hashing to the much faster
wyhash, with a two-level hashing scheme that allows fast scrolling
detection and acceleration by finding a rectangle of scrolled content
and submitting that as a large screen-to-screen copy rather than updated
tiles.

There's three components that together make the experience with screen updates that contain a ton of pixels, common with modern apps on a 4k screen, much better:

- Switch from CRC32 to wyhash. CRC is a fast hash function in hardware but the byte-serial table is slow in software since every byte requires a serial wait for the L1 cache. wyhash takes much better advantage of larger operations and out of order execution. This accelerates a full 4k update on my computer from 123ms in rdpCapture to 38ms
- Lazily convert to YUV only the tiles that actually changed. Some modern apps like Sublime Text don't take care to submit minimal damage rectangles, and thus huge areas are often unchanged, this avoids the color conversion on those areas which is now the slowest part of capture. This cuts capture down to around 11ms with a small amount of change.
- Detect scrolled regions and make them a screen-to-screen copy: Scrolling is probably the most common cause of large updates during normal desktop use. Other remote desktop solutions like Windows RDP and Citrix detect them and optimize them to screen to screen copies. This PR adds that functionality to xrdp by using a two-level hashing scheme that makes detecting a scrolled rectangle on a 4k screen take under 2ms.

## Known issues

- The scrolling detection makes scrolling much nicer but adds tearing around the edges as the screen copy happens before the new tile updates. I prefer the faster scrolling to the minor tearing but opinions may differ.
- Scroll copies can propogate minor tile compression boundary artifacts into more noticeable slight color fringes along solid backgrounds. They're not very noticeable but it makes some compression artifacts more noticeable than before.

Both of these would require the new Progressive RemoteFX codec to fix, for its ability to mark the beginning and end of a frame involving both tile and screen copy updates, and its tweaked RFX compression to reduce boundary artifacts. This seems like a lot of work and I don't plan on doing it. Might be worth adding a build flag to disable scroll detection if during testing this seems like it might annoy some people.

## TODO list (will add items and tick off as suggestions are made):

There's still some things to do, I'm submitting this now to solicit comments on any other changes that would be preferred before merging this patch.

- [ ] I test using it day to day for a while to see if I notice any major issues, I've only briefly done focused testing so far
- [ ] rename "crc" variables to "hash"
- [ ] squash this down to one commit it's just a bajillion commits right now for my personal git convenience.