@metalefty 2020-06-24
Thank you so much for taking care of xorgxrdp performance. It's a big change so it will take some time to review, test, and merge. We might request some changes but we're open. We accept any kind of performance improvements with keeping compatibility, no license violations. I really appreciate your contribution!