@bolkedebruin 2020-06-24
@trishume awesome work! We tried deploying this for testing, but we havent been able to see a difference. This might be due to the fact that we are using guacamole on top of xrdp and are thus dependent on its screen updates as well. We’ll test it without and see if that changes anything.