@trishume 2020-06-24
@bolkedebruin Something I just realized that I may need to improve: I have some hand-tuned constants in the scroll detection code that I forgot to make adaptive to screen size, so it might be that scroll detection is only reliable on 4k displays or at least with large windows on more normal displays. Basically if you're not seeing tearing at the edges of your window when scrolling, then scroll detection isn't activating.

The other improvement also probably will still be better on a 1080p monitor but instead of a ~100ms worst case improvement it'll be more like 25ms or less which is the kind of difference it's hard to notice without a side-by-side comparison, and only helps some situations.