@bolkedebruin 2020-06-25
@trishume I *think* I got it working. The tearing is quite significant. Content becomes unreadable when scrolling. See screenshots from a Windows 2019 client. 

<img width="1401" alt="Screenshot 2020-06-25 at 11 51 46" src="https://user-images.githubusercontent.com/4282712/85697784-64e3d500-b6da-11ea-995d-1cb36d657627.png">
![Uploading Screenshot 2020-06-25 at 11.51.38.png…]()