@trishume 2020-06-25
@bolkedebruin woah that's really weird. I have no idea where the blue comes from, it should just be one part of the screen moving a little before other parts. Definitely something going wrong there.

Also you mention Windows 2019 client but what's that macOS notification in the upper right? If you're double-hopping through some macOS remote desktop thing it's plausible that would exaggerate tearing, although it still doesn't explain the blue.