@trishume 2020-06-25
@bolkedebruin Okay I think I can reproduce that now that I've finally tested it on a Windows client (took me a bit to figure an issue out first), don't know why it happens though, could actually be that screen to screen copies aren't supported in the windows client in RemoteFX mode which would be real sad because I'd need to either give up on scrolling acceleration or do a larger patch that includes changes to xrdp.