@bolkedebruin 2020-06-25
(I was running w2019 in a VM)

Actually, the client needs to support Remote FX which the official Mac client doesn't do. I can only get fast scrolling / low Latency when remote FX works. 

I think screen copies are supported but it copies something wrongly. The blue / green color is from the background that XRDP sends at login. I think that login screen is not rfx accelerated. So could it be a buffer issue maybe?