@trishume 2020-06-25
@bolkedebruin I've been testing with FreeRDP with RFX turned on, in fact my patch doesn't do anything without RemoteFX. Good catch on the original background. My guess is that unlike FreeRDP, the Windows client uses a separate buffer for RemoteFX tiles so when you do a "screen to screen" copy it just copies from the legacy screen buffer over top of the RemoteFX content. When I try it with Windows with pre-filled login information (so no login screen) I just get black scroll rectangles.

I'm not sure if this is fixable without porting to Progressive RemoteFX. It's possible the older RemoteFX has some kind of screen to screen copy that works though, or there's some way to reference the buffer where the actual RemoteFX tiles are and copy from it.