@Nexarian 2020-06-25
I believe Jay was working on the RemoteFX implementation, and I know I am
willing to help out, and many others who love XRDP are probably willing to
help out as well.

I say we bite the bullet and go for it, even if it'll be a little difficult.

On Thu, Jun 25, 2020 at 5:37 PM Tristan Hume <notifications@github.com>
wrote:

> @bolkedebruin <https://github.com/bolkedebruin> I've been testing with
> FreeRDP with RFX turned on, in fact my patch doesn't do anything without
> RemoteFX. Good catch on the original background. My guess is that unlike
> FreeRDP, the Windows client uses a separate buffer for RemoteFX tiles so
> when you do a "screen to screen" copy it just copies from the legacy screen
> buffer over top of the RemoteFX content. When I try it with Windows with
> pre-filled login information (so no login screen) I just get black scroll
> rectangles.
>
> I'm not sure if this is fixable without porting to Progressive RemoteFX.
> It's possible the older RemoteFX has some kind of screen to screen copy
> that works though, or there's some way to reference the buffer where the
> actual RemoteFX tiles are and copy from it.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/neutrinolabs/xorgxrdp/pull/167#issuecomment-649830687>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AA4UBHJYN3XYA7D4CIZZ6I3RYO7SJANCNFSM4OGH64PQ>
> .
>