@jsorg71 2020-06-25
Wow, looks like some interesting bits in here.  Thanks for this.  I also been looking at Adler32 instead of Crc32 mainly because I want to run the calculation in the fragment shader.
I like the idea of doing the crc(hash) first and then you know what further work needs to be done.
I'm interested in the scroll detection too.  We have CopyArea and WindowCopy but they don't help for things like a web browser scrolling.
Unfortunately, I don't think we can mix orders and codecs(surface commands).  You can mix orders with bitmap updates but when in codec mode like RemoteFX, everything has to be RemoteFX.
Same with GFX, you can't mix.  When you start GFX, everything has to be GFX.  GFX has scroll functions so that's good.  I did start a GFX branch in xrdp so lets work together on that.