@bolkedebruin 2020-06-26
@jsorg71 can you share where that GFX branch is?

@trishume won’t it be possible to use rfx tiles (not sure but maybe here https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-rdprfx/89e669ed-b6dd-4591-a267-73a72bc6d84e) search for “tile”. Isn’t RFX progressive “just” a encoding format?