@trishume 2020-06-27
@bolkedebruin I found it here https://github.com/neutrinolabs/xrdp/compare/devel...jsorg71:egfx

@jsorg71 Cool, I'm glad there already looks to be good progress towards implementing EGFX and I hope the scroll detection possibilities provide additional reason behind completing it. I'm still unsure if I'll have the motivation to help out with that but we'll see, maybe I will. This has mostly been a fun side project for me, and it turns out that when testing with a Windows client and my good internet, scrolling already feels pretty good to me without any acceleration so I don't have as strong a demand for that anymore.