@bolkedebruin 2020-07-08
@jsorg71 @trishume @metalefty I'm willing to sponsor some of this. In short: GFX (RemoteFX is deprecated I guess and is not properly supported by the latest Mac client) and online resizing. You can reach me at bolke at apache dot org to work something out.